requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
dj-database-url==2.2.0
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
msgpack==1.0.8
packaging==24.1
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
